0.1.3
Speed up scoring.

Open the alignment file once in each worker process, instead of in each
call to score_feature. There's a surprising amount of overhead in
pysam's opening of BAM files. The actual fetch calls are really quick.

The AlignmentFile instances are not supposed to be safe to share between
processes, so each worker still has to have its own, but it's a big
reduction in overhead; things are roughly twice as fast now.

Also refactor make_cut_matrix, mainly to enable profiling.